move load data code out of management command
