+++ corehq/apps/dump_reload/sql/__init__.py
@@ -1,0 +1,4 @@
+from corehq.apps.dump_reload.sql.dump import dump_sql_data
+from corehq.apps.dump_reload.sql.load import load_sql_data
+
+__all__ = ['dump_sql_data', 'load_sql_data']
